refactor (#16) : 엔티티 & 도메인 모델 병합
