+++ src/main/java/com/limito/limitedproduct/application/service/LimitedProductServiceV1.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
+import com.limito.limitedproduct.domain.model.ProductItem;
 import com.limito.limitedproduct.domain.repository.LimitedProductItemRepository;
-import com.limito.limitedproduct.domain.vo.ProductItem;
 import com.limito.limitedproduct.presentation.dto.request.GetPurchaseAmountLimitRequestV1;
 import com.limito.limitedproduct.presentation.dto.response.GetPurchaseAmountLimitResponseV1;
