+++ src/main/java/com/limito/limitedproduct/domain/repository/LimitedProductItemRepository.java
@@ -3,7 +3,7 @@ package com.limito.limitedproduct.domain.repository;
 import java.util.List;
 import java.util.UUID;
 
-import com.limito.limitedproduct.domain.vo.ProductItem;
+import com.limito.limitedproduct.domain.model.ProductItem;
 
 public interface LimitedProductItemRepository {
