--- src/main/java/com/limito/limitedproduct/domain/vo/Product.java
@@ -1,12 +1,0 @@
-package com.limito.limitedproduct.domain.vo;
-
-import java.util.UUID;
-
-public class Product {
-
-	private UUID id;
-	private UUID categoryId;
-	private String name;
-	private Long sellerId;
-	private String sellerName;
-}
