--- src/main/java/com/limito/limitedproduct/infrastructure/persistence/entity/OptionStatusEnum.java
@@ -1,6 +1,0 @@
-package com.limito.limitedproduct.infrastructure.persistence.entity;
-
-public enum OptionStatusEnum {
-
-	READY, OPEN
-}
