+++ src/main/java/com/limito/limitedproduct/infrastructure/persistence/repository/LimitedProductItemJpaRepository.java
@@ -4,7 +4,7 @@ import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.limito.limitedproduct.infrastructure.persistence.entity.ProductItemEntity;
+import com.limito.limitedproduct.domain.model.ProductItem;
 
-public interface LimitedProductItemJpaRepository extends JpaRepository<ProductItemEntity, UUID> {
+public interface LimitedProductItemJpaRepository extends JpaRepository<ProductItem, UUID> {
 }
