+++ src/main/java/com/limito/limitedproduct/presentation/dto/response/GetPurchaseAmountLimitResponseV1.java
@@ -3,7 +3,7 @@ package com.limito.limitedproduct.presentation.dto.response;
 import java.util.List;
 import java.util.UUID;
 
-import com.limito.limitedproduct.domain.vo.ProductItem;
+import com.limito.limitedproduct.domain.model.ProductItem;
 
 import lombok.AccessLevel;
 import lombok.Builder;
